chore(ci): finish sign refactor and limit pr build
First attempt to copy "avoid subshell, sign image with env. var" to this
repo was incomplete as this workflow has two distinct build jobs and the
second one was missed.

This commit finishes the job, plus prevents the following steps from
running in a PR, again, to match the main repo:
- login to GHCR
- push to GHCR
- sign image
- echo push outputs