chore: add new public key for akmods (#212) (#102)
* chore: add new public key for akmods (#212)

Adds the new public_key.der which will be used for signing akmods in the
future, making it available in the new path today. This will allow users
to start importing it as a new MOK immediately, so they are not caught
off guard when the key switches over in the future.

* docs: add info and sample command for new signing key

Specific signing cutover date of June 17, 2023, 00:00:00 UTC.